move repurpose teams.ts as operations.ts to hold all operations related types
